Refactor Entity out of Particle for use by kt_schema targets (#3700)
- Clean ups for PR

* Refactor Entity out of Particle for use by kt_schema targets

* Add tutorial BUILD
